Auto modes implemented
If ane click LIM, FNL or tag with ctrl pressed they get a green background. In
this mode the corresponding method is called on every hotness change. So if your
images have GPS coordinates set, you can set FNL automode and tag automode. Then
go through your images. asGPS atomatically calls the FNL for every image you select
and tags it with the values. Set the checkboxes to yellow for this, so that asGPS
does not overwrite values already set.