GoogleTab with raw infos and timestamp splitting
New tab with raw information of the google geocode lookus and GPS coordinates
to cut'n' paste with the mouse in other applications and fields.

If the GPS tags include date and time both in the time field, it is possible
now to split them up in the date and time field. There is an option for this
in the settings tab.