Deferred loading and updating, map regions
The loading of a map on startup is now deferred. Also the loading of the config
file.

There is a new option for region codes to optimize the google search.
Map language and region can be arbitrary 2-letter-language codes, not only the
ASP languages.

Project file now with system includes. For the moment in current directory.

JavaScript map files refactorized.