Keep map on hotness change
The map position is now hold on hotness change from the former selected image. This
enables the possibility to mark positions continuously. Also the coordinates lat/lng
are hold. It can be disabled by setting the option

keepMapOnHotnessChange to false

in the config file manually. This is a guru hack, so there is no settings checkbox.

If there is no gps lat/lng tagging in the current hotness the marker in the map is
set to gray.